combine searchSeedMS and searchSeedTW
Passing a function pointer for moveBlob was definitely too costly, but
switching based on an enum seems to be fine.

Note that the order of the enum seems to be important - i got run times
that were 0.2 µs/seed slower the other way around. Probably it's just
affecting code layout and the speedup is coincidental. Possibly it's
just measurement noise (although it seemed pretty consistent.)

MS is about 1.5× slower than TW.